Added group settings for row classes and ignored CSS files
After the groups are formed, we now loop through and determine the row
classes, like `grid`. This can really only be done after processing the
groups.

Also added a check for whether the file type was a markdown or image
file and if it wasn't, we skip it in the processing. This effectively
skips CSS and JS files.